fix: run gofmt, golangci-lint and address all lint issues
- Fix gofmt formatting in 3 files (dlq.go, common_test.go, client_test.go)
- Fix errcheck: add _ = prefix to unchecked Close/Write calls in redis pool,
  test files, gateway handlers
- Fix errorlint: use errors.Is() instead of == for redis.Nil checks in
  session_cache.go and idempotency.go
- Fix govetshadow: rename shadowed err in dlq_test.go and
  circuit_breaker.go (cb → cb2)
- Fix revive: rename EapCode→Code, EapMethod→Method, EapResult→Result,
  EapTlsFlags→TLSFlags, Packet.Id→ID, ExpectedId→ExpectedID across eap/
  package; add missing comments to SessionState*, Result*, NotificationType*,
  AaaRequestDuration, EapRounds, etc.; rename consts in biz/router.go,
  aaa/router.go, aaa/config.go, proto/aaa_transport.go
- Fix unparam: remove unused decision param from sendRADIUS/sendDIAMETER;
  rename unused addr param to _; add nolint for handlePacket underscore param
- Fix unconvert: remove redundant uint8() casts in radius_test.go;
  remove redundant uint8() in diameter eap_avp.go
- Fix gocyclo: extract buildDERMessage helper in diameter/client.go and
  diameter_forward.go to reduce complexity below threshold
- Fix unparam (diameter/client.go): mark vendor param as _ in addAVP closure
- Fix unused parameters and empty branches in diameter_forward.go
- Fix missing comments in NRFDiscoveryCache.Get/Set methods
- Add exclusion rules in .golangci.yml for test files, compose/, stutter
  naming patterns, and API-compatible underscore params

Made-with: Cursor

+++ internal/diameter/eap_avp.go
@@ -61,7 +61,7 @@ func DecodeSnssaiAVP(m *diam.Message) (sst uint8, sd string, err error) {
 	if !ok {
 		// Fallback: try as raw octet string (SST only).
 		if os, ok := avps[0].Data.(datatype.OctetString); ok && len(os) >= 1 {
-			return uint8(os[0]), "", nil
+			return os[0], "", nil
 		}
 		return 0, "", fmt.Errorf("diameter: 3GPP-S-NSSAI AVP has unexpected type %T", avps[0].Data)
 	}
